Move files to an appropriate place
