+++ decidim-admin/app/commands/decidim/admin/remove_admin.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Admin
+    # A command to remove the admin privilege to an user.
+    class RemoveAdmin < Rectify::Command
+      # Public: Initializes the command.
+      #
+      # form - A form object with the params.
+      def initialize(user)
+        @user = user
+      end
+
+      def call
+        return broadcast(:invalid) unless user
+
+        user.update_attributes!(admin: false)
+
+        broadcast(:ok)
+      end
+
+      private
+
+      attr_reader :user
+    end
+  end
+end
